refactor (consensus): create pacemaker submodule
