+++ consensus/debugging.go
@@ -86,31 +86,3 @@ func (m *consensusModule) togglePacemakerManualMode(_ *messaging.DebugMessage) {
 	}
 	m.paceMaker.SetManualMode(newMode)
 }
-
-// This Pacemaker interface is only used for development & debugging purposes.
-type PacemakerDebug interface {
-	SetManualMode(bool)
-	IsManualMode() bool
-	ForceNextView()
-}
-
-type paceMakerDebug struct {
-	manualMode                bool
-	debugTimeBetweenStepsMsec uint64
-
-	// IMPROVE: Consider renaming to `previousRoundQC`
-	quorumCertificate *typesCons.QuorumCertificate
-}
-
-func (p *paceMaker) IsManualMode() bool {
-	return p.manualMode
-}
-
-func (p *paceMaker) SetManualMode(manualMode bool) {
-	p.manualMode = manualMode
-}
-
-func (p *paceMaker) ForceNextView() {
-	lastQC := p.quorumCertificate
-	p.startNextView(lastQC, true)
-}
