+++ consensus/pacemaker/pacemaker_debug.go
@@ -1,0 +1,31 @@
+package pacemaker
+
+import typesCons "github.com/pokt-network/pocket/consensus/types"
+
+// This Pacemaker interface is only used for development & debugging purposes.
+type PacemakerDebug interface {
+	SetManualMode(bool)
+	IsManualMode() bool
+	ForceNextView()
+}
+
+type paceMakerDebug struct {
+	manualMode                bool
+	debugTimeBetweenStepsMsec uint64
+
+	// IMPROVE: Consider renaming to `previousRoundQC`
+	quorumCertificate *typesCons.QuorumCertificate
+}
+
+func (p *paceMaker) IsManualMode() bool {
+	return p.manualMode
+}
+
+func (p *paceMaker) SetManualMode(manualMode bool) {
+	p.manualMode = manualMode
+}
+
+func (p *paceMaker) ForceNextView() {
+	lastQC := p.quorumCertificate
+	p.startNextView(lastQC, true)
+}
